requirements.txt
-i https://pypi.org/simple
aniso8601==9.0.1
asgiref==3.7.2; python_version >= '3.7'
boto3==1.34.47; python_version >= '3.8'
botocore==1.34.47; python_version >= '3.8'
django==4.2.10; python_version >= '3.8'
django-cors-headers==4.3.1; python_version >= '3.8'
django-filter==23.5; python_version >= '3.7'
django-graphql-geojson==0.1.4
django-graphql-jwt==0.4.0; python_version >= '3.6'
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
geographiclib==2.0; python_version >= '3.7'
geopy==2.4.1; python_version >= '3.7'
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3; python_version >= '3.6' and python_version < '4'
graphql-relay==3.2.0; python_version >= '3.6' and python_version < '4'
jmespath==1.0.1; python_version >= '3.7'
promise==2.3
psycopg2-binary==2.9.9; python_version >= '3.7'
pyjwt[crypto]==2.8.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
s3transfer==0.10.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
text-unidecode==1.3
typing-extensions==4.9.0; python_version < '3.11'
urllib3==1.26.18; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
